Create networkOverview module
networkOverviewOutput: output function for the graphD3 output
networkOverviewServer: server function. Takes node and edge data
and creates an overview by creating a simpler graph where each node
represents a cluster. Nodes have the size property which is the
number of genes in the cluster. The edges weight property is the
number of edges from one cluster to another. It then uses graphD3Server
to produce a graph.